// src/types/Offer.ts
export interface IOffer {
  aboutSubtitle: string;
  description: string;
  headerLabelFirst: string;
  headerLabelSecond: string;
  headerLabelThird: string;
  headerTextFirst: string;
  headerTextSecond: string;
  headerTextThird: string;
  id: string;
  isActive: boolean;
  logoPath: string;
  name: string;
  presentationLabel: string;
  presentationPath: string;
}